backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
mangum==0.17.0
transformers==4.35.2
torch==2.1.1
scipy==1.11.4
google-api-python-client==2.108.0
python-dotenv==1.0.0 